move resource class to resource file
